requirements.txt
pyyaml
transformers
accelerate
diffusers
xformers
bitsandbytes
torchvision
Pillow